requirements.txt
appdirs==1.4.4
attrs==25.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
cssselect==1.3.0
fake-useragent==2.1.0
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
lxml==5.3.1
lxml_html_clean==0.4.1
outcome==1.3.0.post0
parse==1.20.2
pyee==11.1.1
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
requests-html==0.10.0
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
urllib3==1.26.20
w3lib==2.3.1
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.21.0
